label done. button and screen folder created in components folder
