+++ src/components/button/index.tsx
@@ -1,0 +1,5 @@
+export default function Button({title}: string){
+    return(
+        
+    )
+}
